Added features folder and shared folders; moved components
