+++ src/features/TodoList/TodoList.jsx
@@ -1,0 +1,26 @@
+{
+  /*extract from TodoList.jsx */
+}
+import TodoListItem from '../TodoList/TodoListItem';
+
+function TodoList({ todoList, onCompleteTodo }) {
+  const filteredTodoList = todoList.filter(
+    (todo) => todo.isCompleted === false
+  );
+
+  return todoList.length === 0 ? (
+    <p>Add a todo above to get started</p>
+  ) : (
+    <ul>
+      {filteredTodoList.map((todo) => (
+        <TodoListItem
+          key={todo.key}
+          todo={todo}
+          onCompleteTodo={onCompleteTodo}
+        />
+      ))}
+    </ul>
+  );
+}
+
+export default TodoList;
